Version 1.0!
FINISHED PROJECT!
(Work may continue when I am less burned out)

CHANGES:
- Columns in main screen dynamically adjust to the width of their field
name. Other components in the column are centered relative to the
column. This economizes screen space significantly.
- Columns in team table do not dynamically adjust, but their width is
limited to that of the longest member name in the table. This means that
this name (or any other) can be shifted to any cell and will still fit.

TO DO: (if I ever come back to this)
- Built-in help file
- Clean up the code. Right now, there is a lot of redundancy and the
"Main" file is over 1000 lines, it can probably be condensed to at least
3/4 of that.
- Improve the sorting algorithm. Team numbers come out imbalanced with
the current one, try to improve it/add a different algorithm for cases
when you only have one or two fields.